Replace all floats from thesis-bench-floats to doubles
The kernel was not exactly the same as the thesis-bench-doubles.
There was use of internal CUDA helper_math calls such as `length`,
`dot`, and the operators for float3.
Also no SimParams object in the thesis-bench-doubles branch,
nor GetMooreBoxIndexes (i.e. still had triple for loop over all 27
boxes).

We should introduces this branch as an another improvement over
the thesis-bench-doubles branch, due to these small changes
(depending on benchmark results).

+++ src/bound_space_op.h
@@ -6,10 +6,10 @@
 namespace bdm {
 
 template <typename TSO>
-void ApplyBoundingBox(TSO* sim_object, float lb, float rb) {
+void ApplyBoundingBox(TSO* sim_object, double lb, double rb) {
   // Need to create a small distance from the positive edge of each dimension;
   // otherwise it will fall out of the boundary of the simulation space
-  float eps = 1e-10;
+  double eps = 1e-10;
   auto pos = sim_object->GetPosition();
   for (int i = 0; i < 3; i++) {
     if (pos[i] < lb) {
